1593812: Pass in controllerUUID at bootstrap
This PR fixes bug 1593812 where Amazon and Azure
cannot bootstrap because of the following error:

  failed to bootstrap model: cannot start bootstrap
  instance: missing controller UUID

Rather than back out the injecting commit, I am
changing the bootstrap command to pass in the
controllerUUID as a BootstrapParam, as I am guessing
that was the intention when that param was added.

The existing code gets the controllerUUID from two
sources (BootstrapParam arg, and the UUID in the
config), and a follow up is needed with Andrew
and/or Ian to determine which should be The Way to
specify controller UUID, rather than having it in
two places.